+++ CHANGELOG.md
@@ -10,7 +10,7 @@ All changes to this project will be documented in this file.
 > - 🏠 [Internal]
 > - 💅 [Polish]
 
-## Unreleased
+## v20.3.0 (2017-06-21)
 
 #### 🚀 New Feature
 - Add group flag to categorize test cases